Move fade to au4
