+++ src/effects/builtin/fade/fadeeffect.h
